// packages/1-prisma-cloud/1-extensions/target/src/__tests__/fixtures/widget-contract/emitted/contract.d.ts
export type StorageHash =
  StorageHashBase<'6d63ff5784d780f7f8858becf843959dccf1e3272eab21ce3396dca9d255bd16'>;
export type ExecutionHash =
  ExecutionHashBase<'eeaf1f4046a71ed6f02d863393397a283adf15b8666139245223083206542e82'>;
export type ProfileHash =
  ProfileHashBase<'3916f444a8a17ad749191acf9e08dad97d1a327b88c2f1d45d12f240296aa8b2'>;

export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Widget: { readonly id: Char<36>; readonly name: CodecTypes['pg/text@1']['output'] };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Widget: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Widget: { readonly id: Char<36>; readonly name: CodecTypes['pg/text@1']['output'] };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Widget: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;

type ContractBase = Omit<
  ContractType<{
    readonly namespaces: {
      readonly public: {
        readonly id: 'public';
        readonly kind: 'postgres-schema';
        readonly entries: {
          readonly table: {
            readonly Widget: {
              columns: {
                readonly id: {
                  readonly nativeType: 'character';
                  readonly codecId: 'sql/char@1';
                  readonly nullable: false;
                  readonly typeParams: { readonly length: 36 };
                };
                readonly name: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
              };
              primaryKey: { readonly columns: readonly ['id'] };
              uniques: readonly [];
              indexes: readonly [];
              foreignKeys: readonly [];
            };
          };
        };
      };
    };
    readonly storageHash: StorageHash;
  }>,
  'roots' | 'domain'
> & {
  readonly target: 'postgres';
  readonly targetFamily: 'sql';
  readonly roots: {
    readonly Widget: { readonly namespace: 'public' & NamespaceId; readonly model: 'Widget' };
  };
  readonly domain: {
    readonly namespaces: {
      readonly public: {
        readonly models: {
          readonly Widget: {
            readonly fields: {
              readonly id: {
                readonly nullable: false;
                readonly type: {
                  readonly kind: 'scalar';
                  readonly codecId: 'sql/char@1';
                  readonly typeParams: { readonly length: 36 };
                };
              };
              readonly name: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: {
              readonly table: 'Widget';
              readonly namespaceId: 'public';
              readonly fields: {
                readonly id: { readonly column: 'id' };
                readonly name: { readonly column: 'name' };
              };
            };
          };
        };
      };
    };
  };
  readonly capabilities: {
    readonly postgres: {
      readonly distinctOn: true;
      readonly jsonAgg: true;
      readonly lateral: true;
      readonly limit: true;
      readonly orderBy: true;
      readonly returning: true;
    };
    readonly sql: {
      readonly defaultInInsert: true;
      readonly enums: true;
      readonly lateral: true;
      readonly returning: true;
      readonly scalarList: true;
    };
  };
  readonly extensions: {};
  readonly execution: {
    readonly executionHash: ExecutionHash;
    readonly mutations: {
      readonly defaults: readonly [
        {
          readonly ref: {
            readonly namespace: 'public';
            readonly table: 'Widget';
            readonly column: 'id';
          };
          readonly onCreate: { readonly kind: 'generator'; readonly id: 'uuidv4' };
        },
      ];
    };
  };
  readonly meta: {};

  readonly profileHash: ProfileHash;
};

export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;